+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/CapturingRenderersFactory.java
@@ -23,6 +23,7 @@ import android.media.MediaCodec;
 import android.media.MediaFormat;
 import android.os.Bundle;
 import android.os.Handler;
+import android.os.PersistableBundle;
 import android.util.SparseArray;
 import android.view.Surface;
 import androidx.annotation.Nullable;
@@ -273,6 +274,12 @@ public class CapturingRenderersFactory implements RenderersFactory, Dumper.Dumpa
     @Override
     public void signalEndOfInputStream() {
       delegate.signalEndOfInputStream();
+    }
+
+    @RequiresApi(26)
+    @Override
+    public PersistableBundle getMetrics() {
+      return delegate.getMetrics();
     }
 
     // Dumpable implementation