+++ RELEASENOTES.md
@@ -18,6 +18,9 @@
         around an issue that occurs on some devices when switching a surface
         from a secure codec to another codec
         (#8696)[https://github.com/google/ExoPlayer/issues/8696].
+    *   Add `MediaCodecAdapter.getMetrics()` to allow users obtain metrics data
+        from `MediaCodec`.
+        ([#9766](https://github.com/google/ExoPlayer/issues/9766)).
 *   Android 12 compatibility:
     *   Upgrade the Cast extension to depend on
         `com.google.android.gms:play-services-cast-framework:20.1.0`. Earlier